webapi/tct-workers-w3c-tests/workers/SharedWorkerName.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Cui, ZhongyouX <zhongyoux.cui@intel.com>

-->

<html>
  <head>
    <title> shared worker name Tests </title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id=log></div>
    <script>
        setup({timeout:500});
        var DATA1 ="self.foo = 1234: 1234";
        var DATA2 ="self.foo: 1234";
        var DATA4="self.foo: undefined";
        var FileName = './support/shared-worker-name.js';

        var t1 = async_test("Test Description: shared worker name. Make sure we can create a shared worker with no name.");
        // Make sure we can create a shared worker with no name.
        var worker = new SharedWorker(FileName);
        worker.port.postMessage("self.foo = 1234: 1234");
        worker.port.onmessage = t1.step_func(function(evt)
        {
            assert_equals(evt.data, DATA1);
            t1.done();
        });
        worker.port.postMessage("start");

        var t2 = async_test("Test Description: shared worker name. Creating a worker with no name should match an existing worker with no name");
        // Creating a worker with no name should match an existing worker with no name
        var worker = new SharedWorker(FileName);
        worker.port.postMessage("self.foo: 1234");
        worker.port.onmessage = t2.step_func(function(evt)
        {
            assert_equals(evt.data, DATA2);
            t2.done();
        });
        worker.port.postMessage("start");

        var t3 = async_test("Test Description: shared worker name. Creating a worker with an empty name should be the same as a worker with no name.");
        // Creating a worker with an empty name should be the same as a worker with no name.
        var worker = new SharedWorker(FileName, '');
        worker.port.postMessage("self.foo: 1234");
        worker.port.onmessage = t3.step_func(function(evt)
        {
            assert_equals(evt.data, DATA2);
            t3.done();
        });
        worker.port.postMessage("start");

        var t4 = async_test("Test Description: shared worker name. Creating a worker with a different name should not be the same as a worker with no name.");
        // Creating a worker with a different name should not be the same as a worker with no name.
        var worker = new SharedWorker(FileName,'name');
        worker.port.postMessage("self.foo: undefined");
        worker.port.onmessage = t4.step_func(function(evt)
        {
            assert_equals(evt.data, DATA4);
            t4.done();
        });
        worker.port.postMessage("start");

        var t5 = async_test("Test Description: shared worker name. Creating a worker for an alternate URL with no name should work.");
        // Creating a worker for an alternate URL with no name should work.
        var worker = new SharedWorker(FileName);
        worker.port.postMessage("self.foo: undefined");
        worker.port.onmessage = t5.step_func(function(evt)
        {
            assert_equals(evt.data, DATA4);
            t5.done();
        });
        worker.port.postMessage("start");

        var t6 = async_test("Test Description: shared worker name. Creating a worker for an alternate URL with empty name should work.");
        // Creating a worker for an alternate URL with empty name should work.
        var worker = new SharedWorker(FileName,'');
        worker.port.postMessage("self.foo: undefined");
        worker.port.onmessage = t6.step_func(function(evt)
        {
            assert_equals(evt.data, DATA4);
            t6.done();
        });
        worker.port.postMessage("start");
    </script>
  </body>
</html>
